+++ fl_sim/federated_algs/clients_selector/clients_selector_factory.py
@@ -1,6 +1,7 @@
 from fl_sim.configuration import Config
 from .best_selector import BestSelector
 from .best_time_expected_selector import BestTimeExpectedSelector
+from .crop_rotation_selector import CropRotationSelector
 from .dynamic_sampling import DynamicSamplingSelector
 from .loss_and_fairness_selector import LossAndFairnessSelector
 from .best_round_time_selector import BestRoundTimeSelector
@@ -28,3 +29,5 @@ class ClientsSelectorFactory:
             return LossAndFairnessSelector(config, status, logger)
         elif selector == "best_time_expected":
             return BestTimeExpectedSelector(config, status, logger)
+        elif selector == "crop_rotation":
+            return CropRotationSelector(config, status, logger)